requirements.txt
# === Core ===
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0.0
requests>=2.31.0

# === Vector Store & Embeddings ===
chromadb>=0.4.22
tiktoken>=0.5.0
rank-bm25>=0.2.2

# === Web UI ===
streamlit>=1.28.0
streamlit-mermaid>=0.0.4

# === Document Generation ===
fpdf2>=2.7.0

# === API Server ===
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

# === Testing ===
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
